+++ src/d4rkAvatarOptimizer.cs
@@ -25,6 +25,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
     public bool DebugShowUnmergableMaterials = true;
     public bool DebugShowUnmergableTextureMaterials = true;
     public bool DebugShowCrunchedTextures = true;
+    public bool DebugShowNonBC5NormalMaps = true;
     public bool DebugShowLockedInMaterials = true;
     public bool DebugShowUnusedComponents = true;
     public bool DebugShowAlwaysDisabledGameObjects = true;