+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Next Version
+### Features
+* Add info text if any normal maps are found that are not using BC5 compression.
+* Add button to automatically convert all normal maps to BC5 compression.
+
 ### Changes
 * The optimized animator controllers now have `(OptimizedCopy)` appended to their name.
 * Added Limitations section to the readme to clarify that avatars might look broken if shaders are disabled. [(more)](https://github.com/d4rkc0d3r/d4rkAvatarOptimizer/issues/17)